requirements.txt
torch==2.2.0 
torchvision==0.17.0 
torchaudio==2.2.0
evaluate
matplotlib
Pillow
pytorch_lightning
sentencepiece
accelerate
scikit-learn
evaluate
wandb
pytesseract
levenshtein
pyyaml
aiohttp
s3fs
fsspec==2023.6.0
nltk
rouge_score
SmilesPE
clearml
opencv-python
albumentations==1.4.2
timm==0.4.12
rdkit==2024.3.5      
rdkit-pypi==2022.9.5
OpenNMT-py
numpy==1.24.4
accelerate==0.21.0